order sales, updated location for sales, floating button scroll anim

+++ app/src/main/java/com/android/yardsale/helpers/GeopointUtils.java
@@ -36,8 +36,8 @@ public class GeopointUtils {
             location.getLatitude();
             location.getLongitude();
 
-            p1 = new LatLng((int) (location.getLatitude()),
-                    (int) (location.getLongitude()));
+            p1 = new LatLng(location.getLatitude(),
+                    location.getLongitude());
 
             return p1;
         } catch (IOException e) {
